backend/requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
cryptography==43.0.3
Django==5.1.2
django-cors-headers==4.6.0
django-encrypted-model-fields==0.6.5
django-environ==0.11.2
django-fernet-fields==0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
pycparser==2.22
PyJWT==2.9.0
python-telegram-bot==21.7
sniffio==1.3.1
sqlparse==0.5.1
